limit gradient colors to 10 colors
This is dont due to a limitation in the GLSL language where you have to
set an array size using a constant value, i.e I can't use the value of a
uniform for this. And since the wave layers I send to the vertex shader
depend on the colors set in the gradient config, and I don't want to go
the shader string manipulation route, I decided to set an upper limit on
the number of wave layers and use an additional layer property: `isSet`
which declares wether a layer is set by the user or an empty (padding)
layer.

In the vertex shader inside the color blending loop, we break out
of the loop on the first isSet == false value.