adds `destory()` method to gradient class
There was a bug that was surfaced due to react dev mode mounting and
remouting stuff on initial render in which the first unmounted instance
had the WebGL program still running. If I resized the window, the first
WebGL instance's attributes whould still be linked to the same
ArrayBuffer, this caused a WebGL error related to insufecient buffer
size to show in the browser console.